fix(StationManagement): 등록된 역이 하나도 없을 때 에러 메세지가 "등록된 노선이 없습니다"로 출력되는 문제 해결
- 이는 역을 삭제할 때 노선에 등록되어있는지 검증하는 과정을 먼저 검증하기 때문에 그렇습니다. 따라서 노선에 등록되어있는지 검증을 거치기 전에 StationRepository에 역이 있는지 검증하는 로직을 컨트롤러에 추가했습니다.

+++ src/main/java/subway/controller/StationManagement.java
@@ -1,6 +1,7 @@
 package subway.controller;
 
 import subway.exception.AttemptToDeleteDependentObjectException;
+import subway.exception.NoneObjectException;
 import subway.menuType.ManagementMenuType;
 import subway.domain.LineRepository;
 import subway.domain.Station;
@@ -14,6 +15,7 @@ import java.util.stream.Collectors;
 
 public class StationManagement {
     private static final String ERROR_CANNOT_REMOVE = "노선에 등록된 역은 삭제할 수 없습니다.";
+    private static final String ERROR_NO_STATIONS = "등록된 역이 없습니다.";
 
     private static StationView stationView = StationView.getInstance();
     private static ManagementMenuType menu;
@@ -47,6 +49,10 @@ public class StationManagement {
 
     private static void deleteStation() {
         String name = stationView.getNameToDelete();
+
+        if (StationRepository.isEmpty()) {
+            throw new NoneObjectException(ERROR_NO_STATIONS);
+        }
         throwExceptionIfItisInLines(name);
 
         StationRepository.deleteStation(name);
